Don't skip "colour1" values.
The rgb.txt file contains many entries of the form "colour1",
which has the same RGB value as "colour". So, initially, we
removed those entries, and just stripped of a trailing '1'
of the colour name in rgb(). However, there a few cases where
"colour1" and "colour" do *not* map to the same RGB value.

So we now keep the mappings for "colour1", and won't strip a
trailing 1 from a colour name in rgb().