Rename and classify colour names.
The raw data now uses separate words for colours, with the parts
capitalized ('Light Goldenrod Yellow' instead of 'lightgoldenrodyellow').
And each colour will have a type ('RGB' by default, else it will be
prepend the colour name, separated from the colour name with a colon).

rgb() takes an optional class name. With the class name, it searches
for the first time a colour name is a match; else it finds the colour
with that class name. For instance, but the 'RGB' and the 'SGI' class
define a colour 'Grey76', but with different RGB values.

    rgb ("grey76", $RGB_HEX)

will return '#c2c2c2', just as

    rgb ("grey76", $RGB_HEX, "RGB")

However,

    rgb ("grey76", $RGB_HEX, "SGI")

will return '#c1c1c1'.

Alternatively, colour names may be prefixed with the class name,
with the class name and colour name separated with a colon.
That is,

    rgb ("grey76", $RGB_HEX, "SGI")

is identical to:

    rgb ("SGI:grey76", $RGB_HEX)